+++ packages/backend/src/database/migrations/20240222164421_add_notification_frequency_to_alerts.ts
@@ -1,0 +1,13 @@
+import { Knex } from 'knex';
+
+export async function up(knex: Knex): Promise<void> {
+    await knex.schema.alterTable('scheduler', (t) => {
+        t.string('notification_frequency').nullable();
+    });
+}
+
+export async function down(knex: Knex): Promise<void> {
+    await knex.schema.alterTable('scheduler', (t) => {
+        t.dropColumn('notification_frequency');
+    });
+}