+++ packages/backend/src/models/CommentModel/CommentModel.ts
@@ -41,6 +41,7 @@ export class CommentModel {
             const structuredComment: Comment = {
                 commentId: comment.comment_id,
                 text: comment.text,
+                textHtml: comment.text_html,
                 replyTo: comment.reply_to ?? undefined,
                 user: { name: `${comment.first_name} ${comment.last_name}` },
                 createdAt: comment.created_at,
@@ -48,6 +49,7 @@ export class CommentModel {
                 replies: [],
                 canRemove:
                     canUserRemoveAnyComment || comment.user_uuid === userUuid,
+                mentions: comment.mentions,
             };
 
             // Directly attach to parent if it's a reply and the parent exists
@@ -144,8 +146,10 @@ export class CommentModel {
         dashboardUuid: string,
         dashboardTileUuid: string,
         text: string,
+        textHtml: string,
         replyTo: string | null,
         user: LightdashUser,
+        mentions: string[],
     ): Promise<string> {
         const { savedChartUuid } =
             await this.checkDashboardTileExistsInDashboard(
@@ -157,10 +161,12 @@ export class CommentModel {
         )
             .insert({
                 text,
+                text_html: textHtml,
                 dashboard_tile_uuid: dashboardTileUuid,
                 reply_to: replyTo ?? null,
                 user_uuid: user.userUuid,
                 saved_chart_uuid: savedChartUuid ?? null,
+                mentions,
             })
             .returning('comment_id');
 
@@ -228,7 +234,13 @@ export class CommentModel {
 
     async getComment(commentId: string) {
         const result = await this.database(DashboardTileCommentsTableName)
-            .select('user_uuid', 'dashboard_tile_uuid', 'reply_to')
+            .select(
+                'user_uuid',
+                'dashboard_tile_uuid',
+                'reply_to',
+                'mentions',
+                'dashboards.dashboard_uuid',
+            )
             .where('comment_id', commentId)
             .first();
 
@@ -238,6 +250,7 @@ export class CommentModel {
             userUuid: result.user_uuid,
             dashboardTileUuid: result.dashboard_tile_uuid,
             replyTo: result.reply_to,
+            mentions: result.mentions,
         };
     }