+++ packages/common/src/types/comments.ts
@@ -1,6 +1,7 @@
 export type Comment = {
     commentId: string;
     text: string;
+    textHtml: string;
     createdAt: Date;
     user: {
         name: string;
@@ -9,6 +10,7 @@ export type Comment = {
     replies?: Comment[];
     resolved: boolean;
     canRemove: boolean;
+    mentions: string[];
 };
 
 export type ApiCommentsResults = Record<string, Comment[]>;